add rentevasteperiodes function

+++ Classes/Person.cs
@@ -8,11 +8,11 @@ namespace Hypo.Classes
 {
     public class Person
     {
-        public decimal brutoJaarSalaris { get; set; }
+        public double brutoJaarSalaris { get; set; }
         public int rentevastePeriode { get; set; }
-        public decimal rentePercentage { get; set; }
+        public double rentePercentage { get; set; }
         public bool heeftPartner { get; set; }
-        public decimal partnerBrutoJaarSalaris { get; set; } = 0;
+        public double partnerBrutoJaarSalaris { get; set; } = 0;
         public bool heeftStudieSchuld { get; set; }
         public int postcode { get; set; }
         public bool magLenen { get; set; }
